# docs/openapi.yaml
openapi: 3.0.0
servers:
  - description: CERCLES API
    url: http://localhost:8080
info:
  version: "1.0.0"
  title: CERCLES API Documentation
  description: >-
    Esta API proporciona los endpoints para la aplicación CERCLES.
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Ingrese su token JWT para autenticar las solicitudes"
security:
  - BearerToken: []
...
paths:
...
  /api/usuarios/{id}:
    get:
      tags:
        - Usuarios
      summary: Obtener información de un usuario por ID
      description: >
        Este endpoint permite obtener la información de un usuario específico mediante su identificador único (ID).
        Devuelve detalles como el nombre, correo y rol del usuario.
      parameters:
        - name: id
          in: path
          required: true
          description: El identificador único del usuario.
          schema:
            type: integer
      responses:
        "200":
          description: Información del usuario obtenida exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: El identificador único del usuario.
                  nombre:
                    type: string
                    description: El nombre del usuario.
                  correo:
                    type: string
                    description: El correo electrónico del usuario.
                  rol:
                    type: string
                    description: El rol del usuario (Estudiante o Profesor).
        "404":
          description: Usuario no encontrado
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no encontrado.
  /api/usuarios/crear:
    post:
      tags:
        - Usuarios
      summary: Crear un nuevo usuario
      description: >
        Este endpoint permite a los profesores crear un nuevo usuario en el sistema. 
        El usuario puede ser de tipo Estudiante o Profesor y requiere un correo único.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                correo:
                  type: string
                  description: El correo electrónico único del usuario a crear.
                nombre:
                  type: string
                  description: El nombre del usuario.
                rol:
                  type: string
                  enum:
                    - Estudiante
                    - Profesor
                  description: El rol del usuario (Estudiante o Profesor).
              required:
                - correo
                - nombre
                - rol
      responses:
        "201":
          description: Usuario creado exitosamente
          content:
            text/plain:
              schema:
                type: string
                example: Usuario creado exitosamente
        "400":
          description: Solicitud inválida (Rol no válido)
          content:
            text/plain:
              schema:
                type: string
                example: Rol inválido
        "403":
          description: Usuario no autenticado
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado
        "409":
          description: Conflicto (Correo ya registrado)
          content:
            text/plain:
              schema:
                type: string
                example: Usuario con el correo ya existe
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al crear el usuario
  /api/usuarios/profesores:
    get:
      tags:
        - Usuarios
      summary: Obtener lista de profesores
      description: >
        Este endpoint permite obtener un listado de todos los profesores registrados en el sistema. 
        Devuelve información básica de cada profesor, como su ID, nombre y correo.
      responses:
        "200":
          description: Lista de profesores obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: El identificador único del profesor.
                    nombre:
                      type: string
                      description: El nombre del profesor.
                    correo:
                      type: string
                      description: El correo electrónico del profesor.
        "403":
          description: Usuario no autenticado
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error al obtener los profesores
  /api/usuarios/datos:
    get:
      tags:
        - Usuarios
      summary: Obtener datos del usuario autenticado
      description: >
        Este endpoint permite obtener los datos del usuario actualmente autenticado en el sistema. 
        Devuelve información básica como el nombre, correo, y los nombres de usuario asociados a GitHub y Taiga.
      responses:
        "200":
          description: Datos del usuario obtenidos exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  nombre:
                    type: string
                    description: El nombre del usuario.
                  correo:
                    type: string
                    description: El correo electrónico del usuario.
                  gitUsername:
                    type: string
                    description: El nombre de usuario asociado a GitHub (puede ser nulo).
                  taigaUsername:
                    type: string
                    description: El nombre de usuario asociado a Taiga (puede ser nulo).
        "403":
          description: Usuario no autenticado
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no autenticado.
        "404":
          description: Usuario no encontrado
          content:
            text/plain:
              schema:
                type: string
                example: Usuario no encontrado.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error interno del servidor.
  /api/usuarios/{id}/equipos:
    get:
      tags:
        - Usuarios
      summary: Obtener los equipos asociados a un usuario
      description: >
        Este endpoint permite obtener una lista de los equipos a los que pertenece un usuario específico. 
        Solo el propio usuario autenticado puede acceder a sus datos.
      parameters:
        - name: id
          in: path
          required: true
          description: El identificador único del usuario cuyos equipos se quieren consultar.
          schema:
            type: integer
      responses:
        "200":
          description: Lista de equipos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: El identificador único del equipo.
                    nombre:
                      type: string
                      description: El nombre del equipo.
                    cursoId:
                      type: integer
                      description: El identificador del curso al que pertenece el equipo.
                    cursoNombre:
                      type: string
                      description: El nombre del curso al que pertenece el equipo.
                    evaluadorId:
                      type: integer
                      description: El identificador del profesor evaluador del curso.
                    cursoActivo:
                      type: boolean
                      description: Indica si el curso está activo.
        "403":
          description: Acceso denegado
          content:
            text/plain:
              schema:
                type: string
                example: Acceso denegado. Solo el usuario autenticado puede consultar sus equipos.
        "400":
          description: Solicitud inválida
          content:
            text/plain:
              schema:
                type: string
                example: Error en los datos proporcionados.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error interno del servidor.
  /api/usuarios/{id}/cursos:
    get:
      tags:
        - Usuarios
      summary: Obtener cursos disponibles para un estudiante
      description: >
        Este endpoint permite obtener una lista de los cursos disponibles para un estudiante donde aún no tiene un equipo asignado.
        Solo el propio estudiante autenticado puede consultar sus cursos.
      parameters:
        - name: id
          in: path
          required: true
          description: El identificador único del estudiante cuyos cursos se quieren consultar.
          schema:
            type: integer
      responses:
        "200":
          description: Lista de cursos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: El identificador único del curso.
                    nombreAsignatura:
                      type: string
                      description: El nombre de la asignatura del curso.
                    añoInicio:
                      type: integer
                      description: El año de inicio del curso.
                    cuatrimestre:
                      type: integer
                      description: El cuatrimestre en el que se imparte el curso.
                    activo:
                      type: boolean
                      description: Indica si el curso está activo.
                    equipos:
                      type: integer
                      description: Número de equipos en el curso (valor fijo a 0 para este endpoint).
                    numeroEstudiantes:
                      type: integer
                      description: Número de estudiantes en el curso (valor fijo a 0 para este endpoint).
                    numeroEstudiantesSinEquipo:
                      type: integer
                      description: Número de estudiantes sin equipo en el curso (valor fijo a 0 para este endpoint).
        "403":
          description: Acceso denegado
          content:
            text/plain:
              schema:
                type: string
                example: Acceso denegado. Solo el estudiante autenticado puede consultar sus cursos.
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example: Error interno del servidor.
...